+++ jgclark.Dashboard/src/react/components/ItemGrid.jsx
@@ -2,7 +2,7 @@
 //--------------------------------------------------------------------------
 // A grid layout for items within a section.
 // Called by ItemGrid component.
-// Last updated 2024-07-09 for v2.0.0
+// Last updated for v2.1.0.a
 //--------------------------------------------------------------------------
 
 import React from 'react'
@@ -10,6 +10,9 @@ import type { TSectionItem, TSection } from '../../types.js'
 import ItemRow from './ItemRow.jsx'
 import { useAppContext } from './AppContext.jsx'
 import { logDebug } from '@helpers/react/reactDev.js'
+
+// Set to true to see some subtle shading of section backgrounds
+const showColoredBackgrounds = false
 
 type Props = {
   items: Array<TSectionItem>,
@@ -25,8 +28,14 @@ function ItemGrid({ items, thisSection }: Props): React$Node {
 
   const visibleItems = tasksToShow.map((item) => <ItemRow key={item.ID} item={item} thisSection={thisSection} />)
 
+  const sectionBackgroundColor = (tasksToShow.length === 0 || tasksToShow[0].itemType === 'itemCongrats')
+    ? `color-mix(in srgb, var(--bg-main-color), green 4%)`
+    : showColoredBackgrounds && thisSection.sectionTitleColorPart
+      ? `color-mix(in srgb, var(--bg-main-color), var(--fg-${thisSection.sectionTitleColorPart}) 4%)`
+      : 'none'
+
   return (
-    <div className="sectionItemsGrid" id={`${thisSection.ID}-Section`}>
+    <div className="sectionItemsGrid" id={`${thisSection.ID}-Section`} style={{ backgroundColor: sectionBackgroundColor }}>
       {visibleItems}
     </div>
   )